# docs/api.yaml
openapi: 3.1.0
info:
  title: Tadpoles API Starter
  version: 1.0.0
  description: >-
    Starter OpenAPI contract for the Tadpoles Spring Boot application.
    The API assumes JWT authentication, Spring Security authorization,
    and role-aware responses for Admin, Auditor, Analyst, Advisor, and Client users.
servers:
  - url: /api
    description: Application API root
...
security:
  - bearerAuth: []
paths:
...
  /clients/{clientId}/subscriptions:
    parameters:
      - $ref: '#/components/parameters/ClientId'
    get:
      tags: [Clients]
      summary: Get subscriptions for a client
      operationId: listClientSubscriptions
      x-required-roles: [ADMIN, AUDITOR, ANALYST, ADVISOR, CLIENT]
      responses:
        '200':
          description: Subscription list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [Clients]
      summary: Subscribe a client to a model portfolio
      operationId: createClientSubscription
      x-required-roles: [ADMIN, ADVISOR]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubscriptionRequest'
      responses:
        '201':
          description: Subscription created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionSummary'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /model-portfolios/{modelPortfolioId}:
    parameters:
      - $ref: '#/components/parameters/ModelPortfolioId'
    get:
      tags: [Portfolios]
      summary: Get model portfolio details
      operationId: getModelPortfolio
      x-required-roles: [ADMIN, AUDITOR, ANALYST, ADVISOR, CLIENT]
      responses:
        '200':
          description: Portfolio summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelPortfolioSummary'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      tags: [Portfolios]
      summary: Update portfolio metadata
      operationId: updateModelPortfolio
      x-required-roles: [ADMIN]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateModelPortfolioRequest'
      responses:
        '200':
          description: Portfolio updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelPortfolioSummary'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /model-portfolios/{modelPortfolioId}/holdings:
    parameters:
      - $ref: '#/components/parameters/ModelPortfolioId'
    get:
      tags: [Portfolios]
      summary: Get target holdings for a model portfolio
      operationId: listModelPortfolioHoldings
      x-required-roles: [ADMIN, AUDITOR, ANALYST, ADVISOR, CLIENT]
      responses:
        '200':
          description: Target holding list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelPortfolioHoldingListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [Portfolios]
      summary: Create a target holding for a model portfolio
      operationId: createModelPortfolioHolding
      x-required-roles: [ADMIN]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateModelPortfolioHoldingRequest'
      responses:
        '201':
          description: Holding created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelPortfolioHoldingSummary'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /model-portfolios/{modelPortfolioId}/holdings/{instrumentId}:
    parameters:
      - $ref: '#/components/parameters/ModelPortfolioId'
      - $ref: '#/components/parameters/InstrumentId'
    get:
      tags: [Portfolios]
      summary: Get a target holding for a model portfolio
      operationId: getModelPortfolioHolding
      x-required-roles: [ADMIN, AUDITOR, ANALYST, ADVISOR, CLIENT]
      responses:
        '200':
          description: Target holding record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelPortfolioHoldingSummary'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      tags: [Portfolios]
      summary: Update a target holding for a model portfolio
      operationId: updateModelPortfolioHolding
      x-required-roles: [ADMIN]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateModelPortfolioHoldingRequest'
      responses:
        '200':
          description: Holding updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelPortfolioHoldingSummary'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /analytics/model-portfolios/{portfolioId}:
    parameters:
      - $ref: '#/components/parameters/PortfolioId'
    get:
      tags: [Analytics]
      summary: Return analytics for a specific model portfolio
      operationId: getPortfolioAnalytics
      x-required-roles: [ADMIN, ANALYST, ADVISOR, CLIENT]
      responses:
        '200':
          description: Portfolio analytics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsSummaryResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ClientId:
      name: clientId
      in: path
      required: true
      schema:
        type: integer
        format: int64
    ModelPortfolioId:
      name: modelPortfolioId
      in: path
      required: true
      schema:
        type: integer
        format: int64
    PortfolioId:
      name: portfolioId
      in: path
      required: true
      schema:
        type: integer
        format: int64
    InstrumentId:
      name: instrumentId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  responses:
    BadRequest:
      description: Invalid or malformed request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Forbidden:
      description: Authenticated user does not have permission to access the resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Conflict:
      description: Duplicate resource or conflicting state
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    TooManyRequests:
      description: Too many requests were submitted in a short time window
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    InternalServerError:
      description: Unexpected server-side failure
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
  schemas:
    ApiErrorResponse:
      type: object
      required: [timestamp, status, error, message, path]
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        error:
          type: string
        message:
          type: string
        path:
          type: string
...
    ModelPortfolioSummary:
      type: object
      properties:
        modelPortfolioId:
          type: integer
          format: int64
        modelName:
          type: string
        description:
          type: string
        isActive:
          type: boolean
...
    UpdateModelPortfolioRequest:
      type: object
      properties:
        modelName:
          type: string
        description:
          type: string
        isActive:
          type: boolean
    ModelPortfolioHoldingSummary:
      type: object
      properties:
        modelPortfolioId:
          type: integer
          format: int64
        instrumentId:
          type: integer
          format: int64
        targetWeightPct:
          type: number
          format: float
    ModelPortfolioHoldingListResponse:
      type: object
      properties:
        holdings:
          type: array
          items:
            $ref: '#/components/schemas/ModelPortfolioHoldingSummary'
    CreateModelPortfolioHoldingRequest:
      type: object
      required: [instrumentId, targetWeightPct]
      properties:
        instrumentId:
          type: integer
          format: int64
        targetWeightPct:
          type: number
          format: float
    UpdateModelPortfolioHoldingRequest:
      type: object
      properties:
        targetWeightPct:
          type: number
          format: float
...
    SubscriptionSummary:
      type: object
      properties:
        subscriptionId:
          type: integer
          format: int64
        clientId:
          type: integer
          format: int64
        modelPortfolioId:
          type: integer
          format: int64
        subscribedDate:
          type: string
          format: date
        endedDate:
          type: string
          format: date
          nullable: true
        status:
          type: string
    SubscriptionListResponse:
      type: object
      properties:
        subscriptions:
          type: array
          items:
            $ref: '#/components/schemas/SubscriptionSummary'
    CreateSubscriptionRequest:
      type: object
      required: [modelPortfolioId, subscribedDate]
      properties:
        modelPortfolioId:
          type: integer
          format: int64
        subscribedDate:
          type: string
          format: date
        approvedByUserId:
          type: integer
          format: int64
          nullable: true
...
    AnalyticsSummaryResponse:
      type: object
      properties:
        scope:
          type: string
          description: Scope of the analytics payload, such as admin, advisor, client, or portfolio
        generatedAt:
          type: string
          format: date-time
        metrics:
          type: object
          additionalProperties: true
        series:
          type: array
          items:
            type: object
            additionalProperties: true
...